// sdk/generated/types/federalSubtype.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

/** §2(11)(A)-(C). */
export enum FederalSubtype {
  OccApprovedNonbank,
  UninsuredNationalBank,
  FederalBranch,
  NotApplicable,
}

export type FederalSubtypeArgs = FederalSubtype;

export function getFederalSubtypeEncoder(): FixedSizeEncoder<FederalSubtypeArgs> {
  return getEnumEncoder(FederalSubtype);
}

export function getFederalSubtypeDecoder(): FixedSizeDecoder<FederalSubtype> {
  return getEnumDecoder(FederalSubtype);
}

export function getFederalSubtypeCodec(): FixedSizeCodec<
  FederalSubtypeArgs,
  FederalSubtype
> {
  return combineCodec(getFederalSubtypeEncoder(), getFederalSubtypeDecoder());
}